refactor - example project
