+++ .idea/modules.xml
@@ -6,6 +6,7 @@
       <module fileurl="file://$PROJECT_DIR$/packages/stream_video/melos_stream_video.iml" filepath="$PROJECT_DIR$/packages/stream_video/melos_stream_video.iml" />
       <module fileurl="file://$PROJECT_DIR$/packages/stream_video_flutter/melos_stream_video_flutter.iml" filepath="$PROJECT_DIR$/packages/stream_video_flutter/melos_stream_video_flutter.iml" />
       <module fileurl="file://$PROJECT_DIR$/packages/stream_video_flutter_core/melos_stream_video_flutter_core.iml" filepath="$PROJECT_DIR$/packages/stream_video_flutter_core/melos_stream_video_flutter_core.iml" />
+      <module fileurl="file://$PROJECT_DIR$/melos_stream_video_flutter.iml" filepath="$PROJECT_DIR$/melos_stream_video_flutter.iml" />
     </modules>
   </component>
 </project>
