+++ packages/stream_video_flutter/example/lib/components/circle_avatar.dart
