+++ packages/stream_video_flutter/example/lib/components/participant.dart
@@ -1,4 +1,3 @@
-
 import 'package:flutter/material.dart';
 
 class OnlineParticipant extends StatelessWidget {
